Cleanup main & use new functions
